requirements.txt
ccxt
pandas
ta
oandapyV20
python-dotenv
